+++ .env.example
@@ -1,1 +1,6 @@
-OPENAI_API_KEY=your_openai_api_key_here
+OPENAI_BASE_URL=
+OPENAI_API_KEY=
+OPENAI_MODEL=
+OPENAI_EMBEDDING_MODEL=
+# en or zh
+LANGUAGE=en
